WeaponAttackZoneSelection
Added/Reworked weapon/dreadnought logic so that now weapons do not have static weaponRanges, and instead have weaponAttackZone options that can be accessed with getWeaponAttackZoneOptions.

This means that when a dreadnought attacks with one of its arm's weapons, it will prompt the player (or AI) to select one of the possible attackZone options that the weapon has to make the attack at. This required adding some (likely temporary) changes to the attackZones of the weapons (Ie Assault cannon hits in different pattern for now)